Provide for scheduler-specific stageout hooks
Check for the presence of stageout hook. If found, execute it. The hook
has to be executed at this point in the script because it's between when
the process completes and the signal traps fire on exit. The
job-specific plugins don't have a way to inject code that particular
interval