Introduce slurm_file_stager service
This is (obviously) in a big need of a cleanup. Hardcoded values need to
be removed and made configurable, the authentication should change from
IP-based to shared-secret, etc...

I currently have it set to start automatically when the condor-ce
service starts, but it probably needs to be made conditionl via some
sort of ExecPre or similar.

Anyway, this is what's working locally